fix: Coerce len() to the target int type on Go (no implicit conv)
- emit_go wraps len() in the sized type; Go has no int->uintN coercion
- Add Go/Kt/Py cross-doc real-compile gates (were substring-only)
- Re-bless Go bytes_equal golden: for i < uint32(len(a))

+++ tests/forge/expected/algorithm_bytes_equal.go
@@ -15,7 +15,7 @@ func BytesEqual(a []byte, b []byte) bool {
         return false;
     }
     var i uint32 = 0
-    for i < len(a) {
+    for i < uint32(len(a)) {
         if (a[i] != b[i]) {
             return false;
         }
